# chore(iast): optimize iast by migrating should_iast_patch function to c [backport 2.21]