--- a/tests/appsec/iast/test_env_var.py
+++ b/tests/appsec/iast/test_env_var.py
@@ -1,11 +1,13 @@
-#!/usr/bin/env python3
 import os
 import subprocess
+import threading
 
 import pytest
 
-from ...utils import _build_env
-from .conftest import CONFIG_SERVER_PORT
+from ddtrace.appsec._constants import IAST
+from ddtrace.appsec._iast._ast import iastpatch
+from tests.appsec.iast.conftest import CONFIG_SERVER_PORT
+from tests.utils import _build_env
 
 
 def _run_python_file(*args, **kwargs):
@@ -184,46 +186,39 @@ def test_env_var_iast_enabled_gevent_patch_all_true(capfd):
     assert "hi" in captured.out
 
 
-def test_A_env_var_iast_modules_to_patch(capfd):
+@pytest.mark.parametrize(
+    "module_name,expected_result",
+    (
+        ("please_patch", iastpatch.ALLOWED_USER_ALLOWLIST),
+        ("please_patch.do_not", iastpatch.DENIED_USER_DENYLIST),
+        ("please_patch.submodule", iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST),
+        ("please_patch.do_not.but_yes", iastpatch.ALLOWED_USER_ALLOWLIST),
+        ("please_patch.do_not.but_yes.sub", iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST),
+        ("also", iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST),
+        ("anything", iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST),
+        ("also.that", iastpatch.ALLOWED_USER_ALLOWLIST),
+        ("also.that.but", iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST),
+        ("also.that.but.not", iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST),
+        ("also.that.but.not.that", iastpatch.DENIED_USER_DENYLIST),
+        ("tests.appsec.iast", iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST),
+        ("tests.appsec.iast.sub", iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST),
+        ("ddtrace.allowed", iastpatch.ALLOWED_USER_ALLOWLIST),
+        ("ddtrace", iastpatch.DENIED_NOT_FOUND),
+        ("ddtrace.sub", iastpatch.DENIED_NOT_FOUND),
+        ("hypothesis", iastpatch.DENIED_NOT_FOUND),
+        ("pytest", iastpatch.DENIED_NOT_FOUND),
+    ),
+)
+def test_env_var_iast_modules_to_patch(module_name, expected_result):
     # type: (...) -> None
-    import gc
-    import sys
-
-    from ddtrace.appsec._constants import IAST
-
-    if "ddtrace.appsec._iast._ast.ast_patching" in sys.modules:
-        del sys.modules["ddtrace.appsec._iast._ast.ast_patching"]
-        gc.collect()
-
     os.environ[IAST.PATCH_MODULES] = IAST.SEP_MODULES.join(
-        ["ddtrace.allowed", "please_patch", "also.that", "please_patch.do_not.but_yes"]
+        ["ddtrace.allowed.", "please_patch.", "also.that.", "please_patch.do_not.but_yes."]
     )
-    os.environ[IAST.DENY_MODULES] = IAST.SEP_MODULES.join(["please_patch.do_not", "also.that.but.not.that"])
-    import ddtrace.appsec._iast._ast.ast_patching as ap
+    os.environ[IAST.DENY_MODULES] = IAST.SEP_MODULES.join(["please_patch.do_not.", "also.that.but.not.that."])
+    iastpatch.build_list_from_env(IAST.PATCH_MODULES)
+    iastpatch.build_list_from_env(IAST.DENY_MODULES)
 
-    for module_name in [
-        "please_patch",
-        "please_patch.submodule",
-        "please_patch.do_not.but_yes",
-        "please_patch.do_not.but_yes.sub",
-        "also",
-        "anything",
-        "also.that",
-        "also.that.but",
-        "also.that.but.not",
-        "tests.appsec.iast",
-        "tests.appsec.iast.sub",
-        "ddtrace.allowed",
-    ]:
-        assert ap._should_iast_patch(module_name), module_name
-
-    for module_name in [
-        "ddtrace",
-        "ddtrace.sub",
-        "hypothesis",
-        "pytest",
-    ]:
-        assert not ap._should_iast_patch(module_name), module_name
+    assert iastpatch.should_iast_patch(module_name) == expected_result, module_name
 
 
 def assert_configure_wrong(monkeypatch, capfd, iast_enabled, env):
@@ -293,3 +288,88 @@ def test_config_over_env_var(_iast_enabled, no_ddtracerun, monkeypatch, capfd):
     assert f"IAST env var: {_iast_enabled.capitalize()}" in captured.out
     assert "hi" in captured.out
     assert "ImportError: IAST not enabled" not in captured.err
+
+
+@pytest.mark.parametrize(
+    "module_name,expected_error",
+    [
+        ("", "Invalid module name"),
+        ("a" * 540, "Module name too long"),
+        ("invalid!module@name", "Invalid characters in module name"),
+        ("module/name", "Invalid characters in module name"),
+        (None, "argument 1 must be str, not None"),
+    ],
+)
+def test_should_iast_patch_invalid_input(module_name, expected_error):
+    with pytest.raises(Exception) as excinfo:
+        iastpatch.should_iast_patch(module_name)
+    assert str(excinfo.value) == expected_error
+
+
+def test_should_iast_patch_empty_lists():
+    os.environ[IAST.PATCH_MODULES] = ""
+    os.environ[IAST.DENY_MODULES] = ""
+    iastpatch.build_list_from_env(IAST.PATCH_MODULES)
+    iastpatch.build_list_from_env(IAST.DENY_MODULES)
+
+    assert iastpatch.should_iast_patch("some.module") == iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST
+
+
+def test_should_iast_patch_max_list_size():
+    large_list = ",".join([f"module{i}" for i in range(1000)])
+    os.environ[IAST.PATCH_MODULES] = large_list
+    iastpatch.build_list_from_env(IAST.PATCH_MODULES)
+    assert iastpatch.should_iast_patch("module1") == iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST
+
+
+@pytest.mark.parametrize(
+    "module_name,allowlist,denylist,expected_result",
+    [
+        ("module.both", "module.both.", "module.both.", iastpatch.ALLOWED_USER_ALLOWLIST),
+        ("parent.child", "parent.child.", "parent.", iastpatch.ALLOWED_USER_ALLOWLIST),
+        ("parent.child", "parent.", "parent.child.", iastpatch.DENIED_USER_DENYLIST),
+        ("django.core", "django.core.", "", iastpatch.ALLOWED_USER_ALLOWLIST),
+    ],
+)
+def test_should_iast_patch_priority_conflicts(module_name, allowlist, denylist, expected_result):
+    os.environ[IAST.PATCH_MODULES] = allowlist
+    os.environ[IAST.DENY_MODULES] = denylist
+    iastpatch.build_list_from_env(IAST.PATCH_MODULES)
+    iastpatch.build_list_from_env(IAST.DENY_MODULES)
+    assert iastpatch.should_iast_patch(module_name) == expected_result
+
+
+@pytest.mark.parametrize(
+    "module_name,expected_result",
+    [
+        ("__main__", iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST),
+        ("__init__", iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST),
+        ("a.b.c.d.e.f", iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST),
+        ("module2.sub1", iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST),
+        ("my_module._private", iastpatch.ALLOWED_FIRST_PARTY_ALLOWLIST),
+    ],
+)
+def test_should_iast_patch_special_modules(module_name, expected_result):
+    assert iastpatch.should_iast_patch(module_name) == expected_result
+
+
+def test_should_iast_patch_cached_packages():
+    result1 = iastpatch.should_iast_patch("new.module")
+    result2 = iastpatch.should_iast_patch("new.module")
+
+    assert result1 == result2
+
+
+def test_should_iast_patch_thread_safety():
+    results = []
+
+    def check_module():
+        results.append(iastpatch.should_iast_patch("concurrent.module"))
+
+    threads = [threading.Thread(target=check_module) for _ in range(10)]
+    for t in threads:
+        t.start()
+    for t in threads:
+        t.join()
+
+    assert len(set(results)) == 1