… ddtrace/appsec/_iast/_ast/ast_patching.py
… ddtrace/appsec/_iast/_ast/iastpatch.c
… ddtrace/appsec/_iast/_stacktrace.pyi
… ddtrace/appsec/_python_info/stdlib/__init__.py
… ddtrace/appsec/_python_info/stdlib/module_names_py310.py
… ddtrace/appsec/_python_info/stdlib/module_names_py311.py
… ddtrace/appsec/_python_info/stdlib/module_names_py312.py
… ddtrace/appsec/_python_info/stdlib/module_names_py36.py
… ddtrace/appsec/_python_info/stdlib/module_names_py37.py
… ddtrace/appsec/_python_info/stdlib/module_names_py38.py
… ddtrace/appsec/_python_info/stdlib/module_names_py39.py